Retrieve golang install from go.dev, use latest version when not specified.
This allows the container to be built with versions of go that
are not included in the package manager.

When GO_VERSION is not specified the latest stable version is retrieved.
Previously the major version defaulted to 1.17, this is still the case but
the Dockerfile no longer needs to be updated when a newer major stable version
is released.